(PC-2113): Add count_all_cancelled_bookings() to booking repository
